fix #1047 and better structure for packaging
